migrate: Move SQLite3 migration code into migrate package
